revert one of my changes that caused a crash in a test.
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@21620 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/avt/DBAtts/MetaData/avtDatabaseMetaData.C
@@ -5194,7 +5194,7 @@ avtDatabaseMetaData::ConvertCSGDomainToBlockAndRegion(const char *const var,
     if (mmd && mmd->meshType == AVT_CSG_MESH)
     {
         const intVector& groupIds = mmd->groupIds;
-        if ((int)groupIds.size() > domainAsVisItSeesIt)
+        if (groupIds.size() > domainAsVisItSeesIt)
         {
             int i, j = groupIds[domainAsVisItSeesIt];
             for (i = domainAsVisItSeesIt; i >= 0 && groupIds[i] == j; i--)
